+++ src/main/java/com/garudanest/service/WaterSharingService.java
@@ -15,21 +15,22 @@ public class WaterSharingService {
         if (flats == null) {
             flats = Collections.emptyList();
         }
+
         double totalHeadCount = flats.stream().mapToDouble(Flat::calculateHeadCount).sum();
-        int totalMembers = flats.stream().mapToInt(flat -> flat.getAdults() + flat.getKids() + flat.getGuests()).sum(); // Changed from getRelatives to getGuests
+        double totalMembers = flats.stream().mapToDouble(flat -> flat.getAdults() + flat.getKids() + flat.getGuests()).sum();
 
         totalMembersHolder.setTotalMembers(totalMembers);
 
         return flats.stream()
-                    .map(flat -> {
-                        double flatHeadCount = flat.calculateHeadCount();
-                        double share = (flatHeadCount / totalHeadCount) * tankerCost;
-                        return new String[]{
-                            flat.getFlatNumber(), // Ensure flat number is included
-                            String.valueOf(flatHeadCount),
-                            String.format("%.2f", share)
-                        };
-                    })
-                    .collect(Collectors.toList());
+                .map(flat -> {
+                    double flatHeadCount = flat.calculateHeadCount();
+                    double share = (flatHeadCount / totalHeadCount) * tankerCost;
+                    return new String[]{
+                        flat.getFlatNumber(),
+                        String.valueOf(flatHeadCount),
+                        String.format("%.2f", share)
+                    };
+                })
+                .collect(Collectors.toList());
     }
 }
